Split worker/auth.ts by responsibility (session, guards, org-role)
worker/auth.ts bundled three unrelated concerns behind one import
path: session/db attachment (withSession), simple signed-in/admin
gates (requireUser, requireAdmin), and org-role authorization
(orgRole, requireOrgRole). fallow flagged it as the top refactoring
target: 8 dependent files all coupled through one file meant any
change there had to be considered against every one of them.

Split into session.ts, guards.ts, and org-role.ts, and repointed each
of the 8 importers (index.ts, routes/{links,qr-logos,domains,admin,
auth,billing,orgs}.ts) at the specific file it actually needs. No
behavior change; all 47 worker integration tests still pass.
