+++ src/worker/routes/orgs.ts
@@ -3,7 +3,8 @@ import { HTTPException } from "hono/http-exception";
 import { eq, and, gte, desc, sql, isNull } from "drizzle-orm";
 import * as schema from "../db/schema";
 import type { AppEnv, DB, Env } from "../env";
-import { requireUser, requireOrgRole, orgRole } from "../auth";
+import { requireUser } from "../guards";
+import { requireOrgRole, orgRole } from "../org-role";
 import { orgPlan, userPlan } from "../plan";
 import { sendEmail } from "../email";
 import { deleteQrLogoMsg, enqueueStorage } from "../storage";
